[Refactor] Move LinearMapping to LinearMapping.swift
